Error while restarting using SIGHUP
Pidfile is removed only when interpreter exists. While restarting using
SIGHUP signal, obviously, interpreter doesn't exit, so pidfile is not
removed. We're fixing this by deleting pidfile in
WorkController._shutdown().
