+++ celery/apps/worker.py
@@ -216,15 +216,13 @@ class Worker(configurated):
         }
 
     def run_worker(self):
-        if self.pidfile:
-            pidlock = platforms.create_pidlock(self.pidfile).acquire()
-            atexit.register(pidlock.release)
         worker = self.WorkController(app=self.app,
                                     hostname=self.hostname,
                                     ready_callback=self.on_consumer_ready,
                                     embed_clockservice=self.embed_clockservice,
                                     autoscale=self.autoscale,
                                     autoreload=self.autoreload,
+                                    pidfile=self.pidfile,
                                     **self.confopts_as_dict())
         self.install_platform_tweaks(worker)
         signals.worker_init.send(sender=worker)
